Avoid warnings.

+++ test/ascii_tag.c
@@ -32,6 +32,7 @@
 #include "tif_config.h"
 
 #include <stdio.h>
+#include <string.h>
 
 #ifdef HAVE_UNISTD_H 
 # include <unistd.h> 
@@ -107,8 +108,8 @@ main(int argc, char **argv)
 	for (i = 0; i < NTAGS; i++) {
 		if (!TIFFSetField(tif, ascii_tags[i].tag,
 				  ascii_tags[i].value)) {
-			fprintf (stderr, "Can't set tag %d.\n",
-				 ascii_tags[i].tag);
+			fprintf(stderr, "Can't set tag %d.\n",
+				(int)ascii_tags[i].tag);
 			goto failure;
 		}
 	}
@@ -141,8 +142,8 @@ main(int argc, char **argv)
 	for (i = 0; i < NTAGS; i++) {
 		if (!TIFFGetField(tif, ascii_tags[i].tag, &value)
 		    && !strcmp(value, ascii_tags[i].value)) {
-			fprintf (stderr, "Can't get tag %d.\n",
-				 ascii_tags[i].tag);
+			fprintf(stderr, "Can't get tag %d.\n",
+				(int)ascii_tags[i].tag);
 			goto failure;
 		}
 	}
